add User.py with newUser and generateUser functions, like the game ones
